Move code to respective components
