+++ platform-operator/controllers/verrazzano/component/opensearch/opensearch_component.go
@@ -5,6 +5,9 @@ package opensearch
 
 import (
 	"fmt"
+	"github.com/verrazzano/verrazzano/pkg/k8sutil"
+	"github.com/verrazzano/verrazzano/platform-operator/internal/config"
+	"path"
 
 	"github.com/verrazzano/verrazzano/pkg/vzcr"
 	vzapi "github.com/verrazzano/verrazzano/platform-operator/apis/verrazzano/v1alpha1"
@@ -107,6 +110,19 @@ func (o opensearchComponent) PreInstall(ctx spi.ComponentContext) error {
 
 // Install OpenSearch component install processing
 func (o opensearchComponent) Install(ctx spi.ComponentContext) error {
+	ctx.Log().Progressf("Component: %s, Creating/Updating OpensearchCluster CR", ComponentName)
+	args, err := common.BuildArgsForOpenSearchCR(ctx)
+	if err != nil {
+		return err
+	}
+	// substitute template values to all files in the directory and apply the resulting YAML
+	filePath := path.Join(config.GetThirdPartyManifestsDir(), "opensearch-operator/opensearch_cluster_cr.yaml")
+	yamlApplier := k8sutil.NewYAMLApplier(ctx.Client(), "")
+	err = yamlApplier.ApplyFT(filePath, args)
+
+	if err != nil {
+		return ctx.Log().ErrorfThrottledNewErr("Failed to substitute template values for OpenSearchCluster CR: %v", err)
+	}
 	return nil
 }
 
@@ -131,6 +147,19 @@ func (o opensearchComponent) PostUninstall(context spi.ComponentContext) error {
 
 // PreUpgrade OpenSearch component pre-upgrade processing
 func (o opensearchComponent) PreUpgrade(ctx spi.ComponentContext) error {
+	ctx.Log().Progressf("Component: %s, Creating/Updating OpensearchCluster CR", ComponentName)
+	args, err := common.BuildArgsForOpenSearchCR(ctx)
+	if err != nil {
+		return err
+	}
+	// substitute template values to all files in the directory and apply the resulting YAML
+	filePath := path.Join(config.GetThirdPartyManifestsDir(), "opensearch-operator/opensearch_cluster_cr.yaml")
+	yamlApplier := k8sutil.NewYAMLApplier(ctx.Client(), "")
+	err = yamlApplier.ApplyFT(filePath, args)
+
+	if err != nil {
+		return ctx.Log().ErrorfThrottledNewErr("Component: %s, Failed to substitute template values for OpenSearchCluster CR: %v", ComponentName, err)
+	}
 	return nil
 }
