+++ platform-operator/controllers/verrazzano/component/opensearch/opensearch_component_test.go
@@ -61,7 +61,8 @@ func TestPreUpgrade(t *testing.T) {
 	// The actual pre-upgrade testing is performed by the underlying unit tests, this just adds coverage
 	// for the Component interface hook
 	config.TestHelmConfigDir = "../../../../helm_config"
-	err := NewComponent().PreUpgrade(spi.NewFakeContext(fake.NewClientBuilder().WithScheme(testScheme).Build(), &vzapi.Verrazzano{}, nil, false))
+	config.TestThirdPartyManifestDir = "../../../../thirdparty/manifests"
+	err := NewComponent().PreUpgrade(spi.NewFakeContext(fake.NewClientBuilder().WithScheme(testScheme).Build(), &vzapi.Verrazzano{}, nil, false, profilesRelativePath))
 	assert.NoError(t, err)
 }
 
@@ -642,11 +643,12 @@ func TestPreInstall(t *testing.T) {
 //	THEN no error is returned
 func TestInstall(t *testing.T) {
 	c := createPreInstallTestClient()
+	config.TestThirdPartyManifestDir = "../../../../thirdparty/manifests"
 	ctx := spi.NewFakeContext(c, &vzapi.Verrazzano{
 		Spec: vzapi.VerrazzanoSpec{
 			Components: dnsComponents,
 		},
-	}, nil, false)
+	}, nil, false, profilesRelativePath)
 	err := NewComponent().Install(ctx)
 	assert.NoError(t, err)
 }
