+++ platform-operator/controllers/verrazzano/component/opensearchoperator/module_integration.go
@@ -8,8 +8,6 @@ import (
 	cmconstants "github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/certmanager/constants"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/common/watch"
 	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/nginx"
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/opensearch"
-	"github.com/verrazzano/verrazzano/platform-operator/controllers/verrazzano/component/opensearchdashboards"
 )
 
 // GetWatchDescriptors returns the list of WatchDescriptors for objects being watched by the component
@@ -21,8 +19,6 @@ func (o opensearchOperatorComponent) GetWatchDescriptors() []controllerspi.Watch
 		}),
 		watch.GetModuleUpdatedWatches([]string{
 			nginx.ComponentName,
-			opensearch.ComponentName,
-			opensearchdashboards.ComponentName,
 		}),
 	)
 }
